moving private channel related intents to separate folder
