+++ lambda/custom/index.js
@@ -249,206 +249,6 @@ const NoIntentHandler = {
 	},
 };
 
-const DeleteGroupIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'DeleteGroupIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.deletegroupname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const headers = await helperFunctions.login(accessToken);
-			const speechText = await helperFunctions.deleteGroup(channelName, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('DELETE_CHANNEL.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const MakeGroupModeratorIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'MakeGroupModeratorIntent';
-	},
-	async handle(handlerInput) {
-		try {
-
-			const userNameData = handlerInput.requestEnvelope.request.intent.slots.groupmoderatorusername.value;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupmoderatorchannelname.value;
-			const userName = helperFunctions.replaceWhitespacesDots(userNameData);
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-			const headers = await helperFunctions.login(accessToken);
-			const userid = await helperFunctions.getUserId(userName, headers);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.addGroupModerator(userName, channelName, userid, roomid, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('MAKE_MODERATOR.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const MakeGroupOwnerIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'MakeGroupOwnerIntent';
-	},
-	async handle(handlerInput) {
-		try {
-
-			const userNameData = handlerInput.requestEnvelope.request.intent.slots.groupownerusername.value;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupownerchannelname.value;
-			const userName = helperFunctions.replaceWhitespacesDots(userNameData);
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-			const headers = await helperFunctions.login(accessToken);
-			const userid = await helperFunctions.getUserId(userName, headers);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.addGroupOwner(userName, channelName, userid, roomid, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('ADD_OWNER.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const PostGroupEmojiMessageIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'PostGroupEmojiMessageIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const messageData = handlerInput.requestEnvelope.request.intent.slots.groupemojimessage.value;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupmessageemojichannelname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-			const emojiData = handlerInput.requestEnvelope.request.intent.slots.groupmessageemojiname.value;
-			const emoji = helperFunctions.emojiTranslateFunc(emojiData);
-			const message = messageData + emoji;
-
-			const headers = await helperFunctions.login(accessToken);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.postGroupMessage(roomid, message, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('POST_MESSAGE.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const GroupLastMessageIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'GroupLastMessageIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.grouplastmessagechannelname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const headers = await helperFunctions.login(accessToken);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.groupLastMessage(channelName, roomid, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('GET_LAST_MESSAGE_FROM_CHANNEL.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const GetGroupUnreadMessagesIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'ReadGroupUnreadsIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupunreadschannelname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const headers = await helperFunctions.login(accessToken);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const unreadCount = await helperFunctions.getGroupUnreadCounter(roomid, headers);
-			const speechText = await helperFunctions.groupUnreadMessages(channelName, roomid, unreadCount, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('GET_UNREAD_MESSAGES_FROM_CHANNEL.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
 const HelpIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
@@ -636,6 +436,15 @@ const { MakeModeratorIntentHandler } = require('./handlers/makeModerator')
 const { CreateGrouplIntentHandler } = require('./handlers/createGroup')
 
 const { PostEmojiDirectMessageIntentHandler } = require('./handlers/postEmojiDirectMessage')
+
+const {
+	DeleteGroupIntentHandler,
+    MakeGroupModeratorIntentHandler,
+    MakeGroupOwnerIntentHandler,
+    PostGroupEmojiMessageIntentHandler,
+    GroupLastMessageIntentHandler,
+    GetGroupUnreadMessagesIntentHandler
+} = require('./handlers/privateChannelIntents')
 
 const skillBuilder = new Jargon.JargonSkillBuilder({ mergeSpeakAndReprompt: true }).installOnto(Alexa.SkillBuilders.standard());
